+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased][unreleased]
+### Feature
+* Added multiple requests in parallel mode
 
 ## [3.0.0-rc.9][2016-08-26]
 ### Bugfix